+++ src/Emitter.ts
@@ -1,0 +1,4 @@
+import {EventStation} from './EventStation';
+
+/** An interface to accommodate objects that extend EventStation */
+export interface Emitter extends EventStation { }