+++ typings/all.d.ts
@@ -1,6 +1,0 @@
-/// <reference path="es6-promise/es6-promise.d.ts" />
-/// <reference path="node/node.d.ts" />
-/// <reference path="es6-promise\es6-promise.d.ts" />
-/// <reference path="node\node.d.ts" />
-/// <reference path="rx\rx-lite.d.ts" />
-/// <reference path="rx\rx.async-lite.d.ts" />